+++ src/components/GroupsTable/index.js
@@ -1,0 +1,3 @@
+import GroupsTable from './GroupsTable';
+
+export default GroupsTable;